fix(iter-1): address expert review blocks
- Remove canonical-equality guards in watcher.rs::set_tree_watched_dirs (was breaking Windows where canonicalize prefixes paths with \\?\); inputs are now canonicalized internally and the canonical form is what's stored/compared. Other guards (is_dir, starts_with(root), 1024 cap, atomic populate-then-commit) retained.
- Split watcher.rs tests into watcher_tests.rs via #[path] include  watcher.rs back under 400-line cap.
- Add accepts_non_canonical_input_via_canonicalization regression test.
- Route IPC errors in useTreeWatcher and useFolderChildren folder-changed listener through @/logger.warn instead of silent catch.
- Update docs/architecture.md rule 12 + rule 3, docs/features/watcher.md, docs/features/navigation.md for new folder-changed event and update_tree_watched_dirs command.

Refs #40
